pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "amlgentex"
version = "0.1.0"
description = "AMLgentex: Mobilizing Data-Driven Research to Combat Money Laundering"
readme = "README.md"
requires-python = ">=3.10,<3.14"
license = {file = "LICENSE"}

dependencies = [
    "networkx==3.4.2",
    "numpy==2.1.0",
    "scipy==1.14.0",
    "pandas==2.2.2",
    "pyarrow>=18.0.0",
    "pyyaml>=6.0",
    "tqdm==4.66.5",
    "scikit-learn==1.5.1",
    "optuna==3.6.1",
    "matplotlib==3.9.2",
    "torch>=2.4.0",
    "torchaudio>=2.4.0",
    "torchvision>=0.19.0",
    "torch_geometric==2.6.1",
    "seaborn==0.13.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "jupyterlab>=4.0.0",
    "ipykernel>=6.29.0",
    "notebook>=7.0.0",
]
viz = [
    "umap-learn==0.5.7",
]
network-explorer = [
    "datashader>=0.16.0",
    "holoviews>=1.19.0",
    "panel>=1.3.0",
    "jupyter_bokeh>=4.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src", "src.*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "unit: Unit tests for individual components",
    "integration: Integration tests for component interactions",
    "e2e: End-to-end tests for full simulation pipeline",
    "spatial: Tests for spatial simulation (graph generation)",
    "temporal: Tests for temporal simulation (time-step execution)",
    "slow: Tests that take longer to run",
    "requires_data: Tests that require parameter files",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
